top-level-spice: Build all of SPICE from top-level
This is a new repository that makes it easier to build all
of SPICE with a single 'make' command. It also optimizes
the build of 'common' and 'protocol', which are built only once
and then used by both spice-gtk and spice-server. This makes
it easier and safer to develop code that impacts these shared
components.

This build system was also designed to be compatible with the
standard build system. You can run ./autogen.sh, and it will
switch to auto-tools. I have not tested that feature in a while.

I have not yet attempted to do anything similar with Meson/Ninja.